+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## MOSuite-Galaxy development version
 
-- The Galaxy XML Synthesizer, originally created by @abombin, is now organized as a Python package: `galaxysynth`.
+- This is the first release of the Galaxy Tools & Workflow configuration for [MOSuite](https://github.com/CCBR/MOSuite).
 - Blueprints for MOSuite are stored in `templates/`.
+- The Galaxy XML Synthesizer, originally created by @abombin, is now organized as a Python package: `galaxysynth`.